move KEY_PSRAM_GUARANTEED to CORE.data top level
